refactoring for new consensus
