+++ src/types/CoinRecord.ts
@@ -1,5 +1,3 @@
-import { RpcResponse } from './rpc';
-
 export interface CoinRecord {
     coin: {
         amount: string;
@@ -10,8 +8,5 @@ export interface CoinRecord {
     confirmed_block_index: number;
     spent: boolean;
     spent_block_index: number;
+    timestamp: string;
 };
-
-export interface CoinResponse extends RpcResponse {
-    coin_records: Array<CoinRecord>;
-}
