+++ src/types/block.ts
@@ -3,51 +3,7 @@ import { FoliageSubBlock } from './FoliageSubBlock';
 import { Proof } from './Proof';
 import { RewardClaim } from './RewardClaim';
 import { SubSlot } from './SubSlot';
-import { Vdf, VdfOutput } from './Vdf';
-
-export interface BlockHeader {
-    data: {
-        additions_root: string;
-        aggregated_signature: string;
-        cost: string;
-        extension_data: string;
-        farmer_rewards_puzzle_hash: string;
-        filter_hash: string;
-        generator_hash: string;
-        height: number;
-        pool_target: {
-            max_height: number;
-            puzzle_hash: string;
-        };
-        prev_header_hash: string;
-        proof_of_space_hash: string;
-        removals_root: string;
-        timestamp: string;
-        total_iters: string;
-        total_transaction_fees: string;
-        weight: string;
-    };
-    plot_signature: string;
-}
-
-export interface ProofOfSpace {
-    challenge_hash: string;
-    plot_public_key: string;
-    pool_public_key: string;
-    proof: string;
-    size: number;
-}
-
-export interface ProofOfTime {
-    challenge_hash: string;
-    number_of_iterations: string;
-    output: {
-        a: string;
-        b: string;
-    };
-    witness: string;
-    witness_type: number;
-}
+import { Vdf } from './Vdf';
 
 export interface Block {
     challenge_chain_ip_proof: Proof;
